Fetch 10 days news when init the database
The original one is get all the news, it will take a long time to finish it.
Now the default is ten days, and you can specify the num. If you want to
get all the news, use 'all'.

Get ten days news(default):

    $ python fetch_data.py

specify the num of days:

    $ python fetch_data.py 20

get all the news:

    $ python fetch_data.py all